Moved hop by hop processing to separate dataplane_fabrid file.
